+++ javascript/node/selenium-webdriver/test/safari_test.js
@@ -17,11 +17,12 @@
 
 'use strict';
 
-const webdriver = require('..'),
-  proxy = require('../proxy'),
-  safari = require('../safari'),
-  assert = require('../testing/assert'),
-  test = require('../lib/test');
+const assert = require('assert');
+
+const proxy = require('../proxy');
+const safari = require('../safari');
+const test = require('../lib/test');
+const webdriver = require('..');
 
 
 describe('safari.Options', function() {
@@ -29,13 +30,13 @@ describe('safari.Options', function() {
     it('returns a new Options instance  if none were defined', function() {
       let options = safari.Options.fromCapabilities(
         new webdriver.Capabilities());
-      assert(options).instanceOf(safari.Options);
+      assert(options instanceof safari.Options);
     });
 
     it('returns the options instance if present', function() {
       let options = new safari.Options().setCleanSession(true),
         caps = options.toCapabilities();
-      assert(safari.Options.fromCapabilities(caps)).equalTo(options);
+      assert.equal(safari.Options.fromCapabilities(caps), options);
     });
 
     it('extracts supported WebDriver capabilities', function() {
@@ -44,7 +45,7 @@ describe('safari.Options', function() {
           .set(webdriver.Capability.PROXY, proxyPrefs);
 
       let options = safari.Options.fromCapabilities(caps);
-      assert(options.proxy_).equalTo(proxyPrefs);
+      assert.equal(options.proxy_, proxyPrefs);
     });
   });
 
@@ -58,15 +59,15 @@ describe('safari.Options', function() {
 
     it('returns a new capabilities object if one is not provided', function() {
       let caps = options.toCapabilities();
-      assert(caps).instanceOf(webdriver.Capabilities);
-      assert(caps.get('browserName')).equalTo('safari');
-      assert(caps.get('safari.options')).equalTo(options);
+      assert(caps instanceof webdriver.Capabilities);
+      assert.equal(caps.get('browserName'), 'safari');
+      assert.equal(caps.get('safari.options'), options);
     });
 
     it('adds to input capabilities object', function() {
       let caps = webdriver.Capabilities.safari();
-      assert(options.toCapabilities(caps)).equalTo(caps);
-      assert(caps.get('safari.options')).equalTo(options);
+      assert.equal(options.toCapabilities(caps), caps);
+      assert.equal(caps.get('safari.options'), options);
     });
 
     it('sets generic driver capabilities', function() {
@@ -75,7 +76,7 @@ describe('safari.Options', function() {
       options.setProxy(proxyPrefs);
 
       let caps = options.toCapabilities();
-      assert(caps.get('proxy')).equalTo(proxyPrefs);
+      assert.equal(caps.get('proxy'), proxyPrefs);
     });
   });
 });
@@ -90,12 +91,11 @@ test.suite(function(env) {
       }
     });
 
-    it('can start safaridriver', function() {
+    it('can start safaridriver', async function() {
       service = new safari.ServiceBuilder().build();
 
-      return service.start().then(function(url) {
-        assert(url).matches(/127\.0\.0\.1/);
-      });
+      let url = await service.start();
+      assert(/127\.0\.0\.1/.test(url), `unexpected url: ${url}`);
     });
   });
 }, {browsers: ['safari']});